prepod_bot/requirements.txt
aiogram==3.2.0
python-dotenv==1.0.1
openai==0.27.10
gspread==5.12.4
oauth2client==4.1.3
reportlab==4.0.8
aiofiles~=23.2.1
httpx~=0.24.0
matplotlib==3.9.0
numpy>=1.26.0